feat: remodel the metrics data structure
Instead of a metric holding both its meta-information
as well as its values, the data model is split in two.
The metric type holds the meta-information and allows
for easy creation of custom metrics, while the MetricDatapoint
class (name subject to later changes) is an instance of a metric
and its value.

I'm not sure that attaching a Metric object instance directly
to the datapoint is such a smart idea though. Leads to
serialization complications.